PAE-1621: add reinstate accreditation journey via shared transition module
Generalises the suspend-accreditation journey into an
accreditation-status-transition module driven by a config map (action ->
target status, confirm-page copy, button). Suspend keeps its URLs;
reinstate adds /reinstate/confirm posting { status: 'approved' } to the
backend status-history endpoint. The overview shows Reinstate on the
Accreditation status row for suspended accreditations (admin.write only).
